requirements.txt
PyYAML==6.0.1
msgspec==0.18.5
